make kakao and naver api 50%

+++ src/main/java/kr/ac/kopo/dao/BookDao.java
@@ -13,6 +13,6 @@ public interface BookDao {
 
 	void delete(String name);
 
-	float total(Pager pager);
+	int total(Pager pager);
 
 }
